+++ controllers/TransferController.php
@@ -7,11 +7,9 @@ use dmitrybtn\cp\SortAction;
 use app\models\CardTransfer;
 
 //*****************************************************************************
-class TransferController extends \dmitrybtn\cp\Controller
+class TransferController extends \dmitrybtn\cp\CrudController
 //*****************************************************************************
 {
-	public $model;
-	
 	//-------------------------------------------------------------------------
 	public function actions()
 	//-------------------------------------------------------------------------
@@ -24,38 +22,21 @@ class TransferController extends \dmitrybtn\cp\Controller
 	}
 
 	//-------------------------------------------------------------------------
-	public function behaviors()
-	//-------------------------------------------------------------------------
-	{
-		return [
-			'verbs' => [
-				'class' => \yii\filters\VerbFilter::className(),
-				'actions' => [
-					'delete' => [YII_ENV_TEST ? 'GET' : 'POST'],
-				],
-			],
-		];
-	}
-
-	//-------------------------------------------------------------------------
-	public function titles()
+	public static function title($actionId, $model = null)
 	//-------------------------------------------------------------------------
 	{
 		return [
 			'create' => 'Добавить остановку',
 			'update' => 'Редактировать остановку',
 			'delete' => 'Удалить',
-		];
+		][$actionId];
 	}
 
 	//-------------------------------------------------------------------------
-	public function breads($actionId)
+	public static function breads($actionId, $model = null)
 	//-------------------------------------------------------------------------
 	{
-		return [
-			['label' => $this->model->card->getTitle(), 'url' => ['/card/view', 'id' => $this->model->card->id]],
-			// ['label' => 'Техкарты', 'url' => ['/card/index']],
-		];
+		return CardController::breads('outer', $model->card);
 	}
 
 
@@ -75,7 +56,7 @@ class TransferController extends \dmitrybtn\cp\Controller
 				return $this->redirect(['/card/view', 'id' => $this->model->id_card]); 
 		}	
 
-		return $this->render('form', ['modCardTransfer' => $this->model, 'returnUrl' => $this->getReferrer(['index'])]);
+		return $this->render('form', ['returnUrl' => $this->getReferrer(['index'])]);
 	}
 
 	//-------------------------------------------------------------------------
@@ -95,26 +76,28 @@ class TransferController extends \dmitrybtn\cp\Controller
 				return $this->redirect($returnUrl);
 		}	
 
-		return $this->render('form', ['modCardTransfer' => $this->model, 'returnUrl' => $returnUrl]);
+		return $this->render('form', [$this->model, 'returnUrl' => $returnUrl]);
 	}
 
 	//-------------------------------------------------------------------------
 	public function actionDelete($id)
 	//-------------------------------------------------------------------------
 	{
-		try {		
+		if (Yii::$app->request->isPost || YII_ENV_TEST) {
+			try {		
 
-			$this->find($id)->delete();
+				$this->find($id)->delete();
+				return Yii::$app->getResponse()->redirect($this->getReferrer(), 302, false);
 
-		} catch (\Exception $e) {
+			} catch (\Exception $e) {
 
-			Yii::$app->session->setFlash('error', $e->getMessage());
-			return $this->goReferrer();
+				Yii::$app->session->setFlash('error', $e->getMessage());
+				return Yii::$app->getResponse()->redirect($this->getReferrer());
 
-		}
-
-		return Yii::$app->getResponse()->redirect($this->getReferrer(), 302, false);
+			}		
+		} throw new \yii\web\MethodNotAllowedHttpException('Неверный формат запроса!');
 	}
+
 
 	//-------------------------------------------------------------------------
 	public function find($id)
